[#8] Add: logger into unit tests
